Expander for local and envp vars
